requirements.txt
fastapi==0.128.8
uvicorn[standard]==0.34.0

# Base de datos
SQLAlchemy==1.4.4
psycopg2-binary==2.9.9

# Variables de entorno
python-dotenv==1.2.2

# Seguridad
argon2-cffi==23.1.0
PyJWT[crypto]==2.10.1

# Validaciones
pydantic==2.11.4
email-validator==2.1.0.post1
python-multipart==0.0.26
pandas==2.2.2
pyasn1>=0.6.3


# Excel
openpyxl==3.1.2
Pillow==12.2.0
xlrd>=2.0.1


# PDF
pdfplumber==0.11.9
passlib[bcrypt]==1.7.4
